Bug 2010200 - Use `MOZ_DISTRIBUTION_ID` as `distribution.id` fallback r=glandium,omc-reviewers,toolkit-telemetry-reviewers,emcminn
Differential Revision: https://phabricator.services.mozilla.com/D278918

UltraBlame original commit: f811761a8acb6cdf3513eb051ec4da64de07baa8

+++ browser/components/asrouter/tests/browser/browser_asrouter_targeting.js
@@ -1254,7 +1254,7 @@ add_task(async function check_isChinaRepack() {
     "should select the message for non China repack users"
   );
 
-  prefDefaultBranch.deleteBranch("");
+  prefDefaultBranch.setCharPref("id", "");
 });
 
 add_task(async function check_userId() {
@@ -1513,22 +1513,29 @@ add_task(async function check_userPrefersReducedMotion() {
 });
 
 add_task(async function test_distributionId() {
+  let expectedDefault = Services.prefs
+    .getDefaultBranch(null)
+    .getCharPref("distribution.id", "");
   is(
     ASRouterTargeting.Environment.distributionId,
-    "",
-    "Should return an empty distribution Id"
+    expectedDefault,
+    "Should return the expected default distribution Id"
   );
 
   Services.prefs.getDefaultBranch(null).setCharPref("distribution.id", "test");
-
   is(
     ASRouterTargeting.Environment.distributionId,
     "test",
     "Should return the correct distribution Id"
   );
-
   
-  Services.prefs.getDefaultBranch(null).deleteBranch("distribution.id");
+  if (expectedDefault) {
+    Services.prefs
+      .getDefaultBranch(null)
+      .setCharPref("distribution.id", expectedDefault);
+  } else {
+    Services.prefs.getDefaultBranch(null).deleteBranch("distribution.id");
+  }
 });
 
 add_task(async function test_fxViewButtonAreaType_default() {
